+++ crap.lua
@@ -1,0 +1,4 @@
+return (...):GetGraphene():MakeImportable {
+	a = 6,
+	b = 5
+}